+++ backend/helpdesk/models/support.py
@@ -45,6 +45,11 @@ class SupportTicket(TimeStampedModel, AuditFieldsModel):
         help_text="True if email has been sent",
     )
     attachments = GenericRelation(FileUpload)
+    migrated_issue_id = models.UUIDField(
+        null=True,
+        blank=True,
+        db_index=True,
+    )
 
     history = HistoricalRecords()